Add NN operations test program
Add an elf file compiled from a proof of concept program
written for the accelerator.
It will be used to demonstrate accelerator memory access
as part of a larger program.

An elf file is provided rather than a source copy or a hex
file as it saves adding many the many supporting files this
program requires to the build system but still allows some
level of examination using the standard toolchain.